New note tests. Pt1
